backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlite-utils==3.35.2
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.2
aiohttp==3.9.1
firecrawl-py==0.1.0
